+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Script d'extraction de la configuration extract-conf.php
 - Action automatique LDAP de synchronisation des utilisateurs #430
 - Script d'installation de fréquence par défaut #425
+- Fonction MemoryCache::FlushAll() permettant de vider le cache
 
 # [2.0.7] - 2018-07-18